007_holographic_custom_names : Moved duplicate methods to new HolographUtil
